feat: admin log in and portal are now the same component
